Structure: Move Templates to src subdirectory
Update namespaces as needed.

Also add FieldInterface to output
