+++ src/Templates/Generic.php
@@ -1,15 +1,17 @@
 <?php
 
-namespace ModularWP\Formalize;
+namespace ModularWP\Formalize\Templates;
 
-class DefaultTemplate extends BaseTemplate {
+use ModularWP\Formalize\FieldInterface;
+
+class Generic extends Base {
 
 	/**
 	 * Outputs the input field
 	 *
 	 * @param array $instance
 	 */
-	public function output( $field, $instance = array() ) {
+	public function output( FieldInterface $field, $instance = array() ) {
 
 		// Sets up the output variable.
 		$output = '';
